added set likes function and moved db handling to separate file
