+++ src/components/RecipesSearchGeneral/RecipeComponent.vue
@@ -10,7 +10,7 @@
       <b-card-text>
         <p class="information">Ready in minutes: {{readyInMinutes}}</p>
         <p class="information">Servings: {{servings}}</p>
-        <p class="information">Likes: {{likes}}</p>
+        <p class="information">Likes: {{likes}} <button v-on:click="setLikes(id, likes+1)">Add Like</button></p>
       </b-card-text>
       <b-button href="#" variant="primary" style="color: aliceblue">Go to recipe</b-button>
     </b-card>
@@ -18,13 +18,14 @@
 </template>
 
 <script>
-  import {baseRecipiesApiPhotosUrl, insertNewRow, getLikes, dbTableUrl, dbCorsKey} from '../../shared/constants';
-  import axios from 'axios';
+  import {baseRecipiesApiPhotosUrl} from '../../shared/constants';
+  import {getLikes, setLikes} from '../../shared/DBHandling';
 
   export default {
     data() {
       return {
-        results: []
+        results: [],
+        likes: 0
       }
     },
     props: {
@@ -34,7 +35,6 @@
       servings: Number,
       image: String,
       imageUrls: Array,
-      likes: Number,
     },
     computed: {
       imageUrl: function () {
@@ -42,27 +42,18 @@
       }
     },
     mounted() {
-      let url = new URL(dbTableUrl + `?q={"recipe_id": ${this.id}}`);
-      let self = this
-      let headers = {
-          "content-type": "application/json",
-          "x-apikey": dbCorsKey,
-          "cache-control": "no-cache",
-      }
-      axios.get(url=url, {headers: headers})
-      .then(function(response){
-        if(response.data.length == 0){
-          self.likes = 0;
-        }else{
-          self.likes = response.data[0].likes;
-        }
-      })
+      getLikes(this.id)
+      .then(likes => this.likes = likes);
     },
     created() {
     },
     methods: {
       print: function () {
         console.log(this.id)
+      },
+      setLikes: function(id, likes){
+        setLikes(id, likes)
+        .then(likes => this.likes = likes);
       }
     }
   }
